Add function to assoc-in objects
Defines `lt.object/assoc-in!`, a function that allows the caller to
`assoc-in` objects conforming to the BOT model.

The function takes an object, a sequence of keys, and the value to be
set. If the key already exists in the nested associative structure
described by the sequence, its value is reset; if it doesn't, the new
key-value pair is created, fully reproducing the sequence as hash-maps
if necessary.

Useful to update and create deeply-nested keyvals in a concise fashion,
i.e. without passing full maps as required by `object/merge!`.