Try a default port for socket.io first
Try to use a hard-coded port, 5678, and only ask
for a random free port when it is not available.
Users that have only one connection thus get by
most of the time without having to change the port
in their HTML.

I think it is also better to log errors than to
swallow them.