Fix connector failing to connect
When attempting to connect to a Light Table browser instance within a
brand new mies (David Nolen's bare bones cljs project template) project,
Light Table fails to connect.

It appears that two behaviors with the same trigger :selected are
executed out of order and the target object is destroyed before the
second behavior has a change to execute. This commit moves the behaviors
into one and formats the code.