Add api doc generation through codox
Generate API documentation for namespaces of fns that
were documented in f94bf3af0d382424c69161d5b01680d0ea657a38.
Fns that are meant to be private were marked as :private.
There are a number of public fns in object, files and editor namespaces
that should be documented later