Change delete behavior to be less confusing and time consuming
Currently for each tab whose file is deleted, we pop up a warning
modal. When switching git branches, this is confusing and time consuming
to users as they may sometimes need to deal with several modals.
The new behavior mimics Atom which is reasonable: delete the tab
if the deleted file is unchange and leave the tab alone if the file
has been modified. check-mtime was modified to allow the user to save
a modified tab for a deleted file